requirements.txt
playwright>=1.40.0
browser-cookie3>=0.19.0
Pillow>=10.0.0
